+++ include/openpfc/kernel/simulation/initial_conditions/seed_grid.hpp
@@ -40,6 +40,7 @@
 #include <openpfc/kernel/field/operations.hpp>
 #include <openpfc/kernel/simulation/field_modifier.hpp>
 #include <openpfc/kernel/simulation/initial_conditions/seed.hpp>
+#include <openpfc/kernel/simulation/model.hpp>
 #include <openpfc/kernel/utils/logging.hpp>
 
 namespace pfc {
@@ -119,7 +120,9 @@ public:
       }
     }
 
-    pfc::field::apply(m, get_field_name(), [seeds](const pfc::Real3 &X) {
+    pfc::field::apply(pfc::get_real_field(m, get_field_name()),
+                      pfc::get_world(m), pfc::get_fft(m),
+                      [seeds](const pfc::Real3 &X) {
       for (const auto &seed : seeds) {
         if (seed.is_inside(X)) {
           return seed.get_value(X);